[string] Shrink String.Index to 2 words.
String.Index is 3 words in size, which means that Range<String.Index>
is 6, and Substring is 8 words total. This is pretty wasteful, so make
a very minor adjustment to the index cache's UTF-8 buffer to bring it
down to 2 words total.

+++ stdlib/public/core/StringIndex.swift
@@ -18,7 +18,7 @@ extension String {
     @usableFromInline
     internal var _cache: _Cache
 
-    internal typealias _UTF8Buffer = _ValidUTF8Buffer<UInt64>
+    internal typealias _UTF8Buffer = _ValidUTF8Buffer<UInt32>
     @_frozen // FIXME(sil-serialize-all)
     @usableFromInline
     internal enum _Cache {
